Rename primary and secondary files to match default export name, export new COLORS_PROP_TYPES for aliased values
