+++ src/constants/themes/index.ts
@@ -1,2 +1,2 @@
-export { default as primaryTheme } from './primary';
-export { default as secondaryTheme } from './secondary';
+export { default as primaryTheme } from './primaryTheme';
+export { default as secondaryTheme } from './secondaryTheme';
